requirements.txt
crewai==0.63.6
langchain-openai==0.1.23
streamlit
python-dotenv
duckdb
pandas
langchain-core<0.3.0
langgraph<0.3.0
sentence-transformers==2.6.1
torch>=2.0